cleanup comments, code style, and unit tests
